+++ CHANGELOG.md
@@ -1,1 +1,29 @@
-- added gitlab ci automated build
+## Dandelion Version 0.1.0-dev
+
+this version is the first public release of dandelion server software
+
+# Classic protocol support
+- full
+
+
+# CPE support
+- EnvColors
+- EnvWeatherType
+- EnvMapAspect
+- CustomBlocks
+
+# Commands
+- ban
+- clients
+- help
+- kick
+- level
+- online
+- permission
+- playerInfo
+- plugin
+- say
+- sayRaw
+- serverInfo
+- stop
+- unban